test(resolve-from): ♻️ extract hook into shared module and fix test isolation
- Centralize ESM resolve hook code into src/resolve-from.ts, eliminating
  three-way copy drift between bin/executor.ts, the test fixture, and
  the unit test
- Fixture and executor now call the same registerResolveFromHook()
- Unit test imports resolveFromHookCode from the production module
- Integration test uses mkdtempSync for unique temp dirs per run
- Fix false-positive regex guard in unit test by rewording hook comment
